summary
Ajax functuon added to get data to plotMap.
Function loadmarkers plots the data return by the plotMap function and plots it on the map.
If statement added to check whether search result is empty or not.
plotMap url added.
Views scrapes the address of the restaurant and before saving geocodes the address to convert to coordinates.
Added modules django_google_maps.